S2 VID2: Add wpt_theme_styles() to functions.php
wpt_ is a namespace to give a unique name to all of our fns so that they don't get overriden or clash with plugins
wp_enqueue_style() lets us link to a stylesheet for our particular theme
the first parameter is a unique name (handle) to identify this style sheet - in this case, 'foundation_css'
the second param is a link to this file, and the link is obtaind using get_template_directory_uri(), to which we concatenate the '/css/foundation.css'
we can them add another wp_enqueue_style() to add 'normalize.css'
we can add Google fonts on the third line of the wpt_theme_styles, instead of using @import - note that now, the 2nd param is a link to the google fonts site
But simply writing this function will not generate the files that we need for this to work. We have to do another step here:
use the add_action() function to tell WP when to enqueue those styles.
the 1st param of add_action() is wp_enqueue_scripts - it's a hook that WordPress has that lets us tell it what css/js files to load for a given page
the 2nd param is, again, the location, this time the location is the wpt_theme_styles() function
HOWEVER, this still won't work. We need to add some more code to index, header, and footer files to see this working
We'll take a small detour before that and using the same code, enqueue any JS files we need