Add secondary column below the loop in page-sidebar-left.php
It can be tricky porting over static code, especially when there is a lot.
What lines to copy? Shoud I include this line or not? The answer is,
everything that is a container in the static design, goes outside the loop
in the WordPress file. Everything that is a repeated element, goes inside the loop.
===================================================================================