Fixed errors in closing divs in page-sidebar-left
This part of the tut is not really done the best way, so I wont
mess with it too long. It's a waste of time to set up the sidebar
properly, but it's not important either. I get the main idea, and
that is that you can tweak the Template page to your liking, give
it a comment title at the top of the file and adjust it with html
and css the way you like, and you'll have the option to pick that
design when working on Pages in the admin area, so it's all good.
=================================================================