Controlling navigation using wp_nav_menu()
1st you must add the add_theme_support() fn, and pass it the param of 'menus'
the reason is that by default, WP will not automatically add menus to your theme
after you add the theme support, go to WP admin and create a new menu, call it
Main Menu, add the basic pages we made, save it;
next, in fns.php, tell WP that we have a specific menu set up - we'll tell WP
that this menu exists. We'll do it by using the function register_theme_menus()
Inside it, register_nav_menus() { array ( // menus we want to have on the site ) }
Finally, use add_action ( 'init', ... ) to tell WP where to hook into this, in our
case, when WP is first initializing
Now WP knows there's a primary menu (it exists), but it doesn't know where it is.
We're gonna do that in the next commit, when we come into our header. php file and
replace all of the static navigation code with a simple wp_nav function