Make single-portfolio.php template page
(1) Open page-sidebar-left, save as single-portfolio;
(2) You don't even need the template comment up top, WP will
know that this is the template for displaying custom post types.
(3) We're gpnna have to space out our loop php markup, to get the
result we want
(4) The main content in the loop we'll call in is the custom post
field - the_field(), and pass it the arg of 'images'
(5) Down below, replace <h2>Sidebar..., with <h1><?php the_title()...
(6) Use the_field() tag again - note it's specific to just the ACF
plugin that works with creating custom fields. And this is how you call
any of the custom fields you set: the_field('description')
(7) At the end, add the endwhile; and endif; ?>
=======================================================================