Change the name of front-page to -front-page
(1) if we have a static front page it should use front-page.php
(2) if we're just listing our blog posts it should go straight to the home.php
(3) rename front-page to -front-page so it won't recognize it, and it will now use our home.php instead
(4) 1 way around this is that instead of using the front-page template for everything, what some
theme developers do instead is use the custom templates where you would set that you wnat to have a
static homepage. But then you would select which specific home page or template you want to be using
and they may bypass entirely the frontpage.php which could help them with some flexibility on their site.
=========================================================================================================